// ui/libs/core/src/lib/shared/models/external-dataservice.model.ts
export enum ServiceCategory {
  SCA = 'SCA',
  SourceRepos = 'SourceRepos',
  WorkTrackingTools = 'WorkTrackingTools',
  ContinuousIntegration = 'ContinuousIntegration',
  Performance = 'Performance',
  SAST = 'SAST',
}
// ...
export enum ServiceStatus {
  Mapped = 'mapped',
  NotConnected = 'not-connected',
  Connected = 'connected',
  Disabled = 'disabled',
}

export enum ExternalService {
  AzureBilling = 'AzureBilling',
  Azure = 'Azure',
  Basecamp = 'Basecamp',
  Harvest = 'Harvest',
  NewRelic = 'NewRelic',
  SLABAzure = 'SLABAzure',
  VisualStudioOnline = 'VisualStudioOnline',
  JIRA = 'JIRA',
  Slack = 'Slack',
  GitHub = 'GitHub',
  Sonar = 'Sonar',
  Jenkins = 'Jenkins',
  ApplicationInsights = 'ApplicationInsights',
  Bitbucket = 'Bitbucket',
  BitbucketServer = 'BitbucketServer',
  TFS = 'TFS',
  Asana = 'Asana',
  AWS = 'AWS',
  Nuget = 'Nuget',
  Maven = 'Maven',
  Npm = 'Npm',
  TeamCity = 'TeamCity',
  NVDVulnerabilities = 'NVDVulnerabilities',
  DocioApi = 'DocioApi',
  DocioScanner = 'DocioScanner',
  GitLab = 'GitLab',
  Bamboo = 'Bamboo',
  React = 'React',
  ReactJ = 'React javascript',
  ReactT = 'React typescript',
  Angular = 'Angular',
  Java = 'Java',
  Phyton = 'Phyton',
  TravisCI = 'TravisCI',
  CircleCI = 'CircleCI',
  SonarQube = 'SonarQube',
  SecureCodeWarrior = 'SecureCodeWarrior',
}